remove execution logs and feature compatibility checks
