+++ app/src/components/features/rules/TestThisRule/index.ts
@@ -1,3 +1,3 @@
 export { TestThisRule } from "./TestThisRule";
 export * from "./utils/testReports";
-export { InitImplicitWidgetConfigHandler } from "./hooks/useInitImplicitWidgetConfig";
+export { InitImplicitWidgetConfigHandler } from "./hooks/InitImplicitWidgetConfigHandler";
