+++ app/src/config/constants/compatibility.js
@@ -88,4 +88,8 @@ export const FEATURE_COMPATIBLE_VERSION = {
     [GLOBAL_CONSTANTS.APP_MODES.DESKTOP]: null,
     [GLOBAL_CONSTANTS.APP_MODES.EXTENSION]: "24.2.17",
   },
+  [FEATURES.IMPLICIT_TEST_THIS_RULE]: {
+    [GLOBAL_CONSTANTS.APP_MODES.DESKTOP]: null,
+    [GLOBAL_CONSTANTS.APP_MODES.EXTENSION]: "24.4.18",
+  },
 };
