+++ app/src/config/constants/sub/features.js
@@ -30,6 +30,7 @@ FEATURES.ASYNC_MODIFY_REQUEST_BODY = "ASYNC_MODIFY_REQUEST_BODY";
 FEATURES.SERVE_RESPONSE_WITHOUT_REQUEST = "SERVE_RESPONSE_WITHOUT_REQUEST";
 FEATURES.TEST_THIS_RULE = "test_this_rule";
 FEATURES.TEST_THIS_RULE_SESSION = "test_this_rule_session";
+FEATURES.IMPLICIT_TEST_THIS_RULE = "implicit_test_this_rule";
 FEATURES.DESKTOP_APP_TERMINAL_PROXY = "DESKTOP_APP_TERMINAL_PROXY";
 FEATURES.DESKTOP_APP_SSL_PROXYING = "DESKTOP_APP_SSL_PROXYING";
 FEATURES.DESKTOP_APP_TRAFFIC_TABLE = "DESKTOP_APP_TRAFFIC_TABLE";
